Added a cast from int to machine type

+++ rust.rs
@@ -1,6 +1,7 @@
 #[doc = "Rust wrappers around the raw JS apis"];
 
 import bg = jsapi::bindgen;
+import core::libc::types::os::arch::c95::size_t;
 
 export rt;
 export cx;
@@ -22,7 +23,7 @@ fn rt() -> rt {
 
 impl methods for rt {
     fn cx() -> cx {
-        @cx_rsrc({ptr: JS_NewContext(self.ptr, default_stacksize),
+        @cx_rsrc({ptr: JS_NewContext(self.ptr, default_stacksize as size_t),
                   rt: self})
     }
 }
